+++ src/rerun/extension/ReRunParameter.java
@@ -1,9 +1,5 @@
 package rerun.extension;
 
-import org.springframework.util.StringUtils;
-
-import java.util.StringTokenizer;
-
 /**
  * Created by alutman on 10-Feb-15.
  */
@@ -24,19 +20,4 @@ public class ReRunParameter {
         return value;
     }
 
-    public int height() {
-        int newLines = StringUtils.countOccurrencesOf(value, "\n");
-        int plusOne = 1;
-
-        int wraps = 0;
-        StringTokenizer st = new StringTokenizer(value, "\n");
-        while(st.hasMoreTokens()) {
-            String next = st.nextToken();
-            while(next.length() > 100) {
-                wraps++;
-                next = next.substring(100, next.length());
-            }
-        }
-        return newLines + wraps + plusOne;
-    }
 }